requirements.txt
open_clip_torch==2.23.0
transformers==4.35.2
torch
torchvision
torchaudio
timm
albumentations
opencv-python
scikit-learn
matplotlib
seaborn
pandas
scipy
tensorboard
plotly
numpy
pillow